requirements.txt
asgiref==3.7.2
Django==4.2.2
Pillow==9.5.0
sqlparse==0.4.4

djangorestframework~=3.14.0
celery~=5.3.1